+++ EssentialApp/EssentialApp/SceneDelegate.swift
@@ -88,77 +88,3 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
             .fallback(to: localFeedLoader.loadPublisher)
     }
 }
-
-public extension FeedLoader {
-    typealias Publisher = AnyPublisher<[FeedImage], Error>
-    
-    func loadPublisher() -> Publisher {
-        Deferred {
-            Future(self.load)
-        }.eraseToAnyPublisher()
-    }
-}
-
-extension Publisher where Output == [FeedImage] {
-    func caching(to cache: FeedCache) -> AnyPublisher<Output, Failure> {
-        handleEvents(receiveOutput: {
-            feed in cache.saveIgnoringCompletion(feed: feed) }
-        )
-        .eraseToAnyPublisher()
-    }
-}
-
-extension FeedCache {
-    func saveIgnoringCompletion(feed: [FeedImage]) {
-        self.save(feed) { _ in }
-    }
-}
-
-extension Publisher {
-    func fallback(to fallbackPublisher: @escaping () -> AnyPublisher<Output, Failure>) -> AnyPublisher<Output, Failure> {
-        self.catch { _ in fallbackPublisher() }.eraseToAnyPublisher()
-    }
-}
-
-extension Publisher {
-    func dispatchOnMainQueueIfNeeded() -> AnyPublisher<Output, Failure> {
-        receive(on: DispatchQueue.immediateWhenOnMainQueue).eraseToAnyPublisher()
-    }
-}
-
-extension DispatchQueue {
-    static var immediateWhenOnMainQueue: ImmediateWhenOnMainQueueScheduler {
-        ImmediateWhenOnMainQueueScheduler()
-    }
-    
-    struct ImmediateWhenOnMainQueueScheduler: Scheduler {
-        typealias SchedulerTimeType = DispatchQueue.SchedulerTimeType
-        typealias SchedulerOptions = DispatchQueue.SchedulerOptions
-
-        var now: Self.SchedulerTimeType {
-            DispatchQueue.main.now
-        }
-
-        var minimumTolerance: Self.SchedulerTimeType.Stride {
-            DispatchQueue.main.minimumTolerance
-        }
-
-        func schedule(options: Self.SchedulerOptions?, _ action: @escaping () -> Void) {
-            guard Thread.isMainThread else {
-                return DispatchQueue.main.schedule(options: options, action)
-            }
-            
-            action()
-        }
-
-        /// Performs the action at some time after the specified date.
-        func schedule(after date: Self.SchedulerTimeType, tolerance: Self.SchedulerTimeType.Stride, options: Self.SchedulerOptions?, _ action: @escaping () -> Void) {
-            DispatchQueue.main.schedule(after: date, tolerance: tolerance, options: options, action)
-        }
-
-        /// Performs the action at some time after the specified date, at the specified frequency, optionally taking into account tolerance if possible.
-        func schedule(after date: Self.SchedulerTimeType, interval: Self.SchedulerTimeType.Stride, tolerance: Self.SchedulerTimeType.Stride, options: Self.SchedulerOptions?, _ action: @escaping () -> Void) -> any Cancellable {
-            DispatchQueue.main.schedule(after: date, interval: interval, tolerance: tolerance, options: options, action)
-        }
-    }
-}
